Collect record component annotations and move record processing code to its own file
