+++ src/org/jetbrains/java/decompiler/main/ClassWriter.java
@@ -11,7 +11,6 @@ import org.jetbrains.java.decompiler.main.rels.ClassWrapper;
 import org.jetbrains.java.decompiler.main.rels.MethodWrapper;
 import org.jetbrains.java.decompiler.modules.decompiler.ExprProcessor;
 import org.jetbrains.java.decompiler.modules.decompiler.exps.*;
-import org.jetbrains.java.decompiler.modules.decompiler.stats.BasicBlockStatement;
 import org.jetbrains.java.decompiler.modules.decompiler.stats.RootStatement;
 import org.jetbrains.java.decompiler.modules.decompiler.stats.Statement;
 import org.jetbrains.java.decompiler.modules.decompiler.vars.VarTypeProcessor;
@@ -19,7 +18,6 @@ import org.jetbrains.java.decompiler.modules.decompiler.vars.VarVersionPair;
 import org.jetbrains.java.decompiler.modules.renamer.PoolInterceptor;
 import org.jetbrains.java.decompiler.struct.*;
 import org.jetbrains.java.decompiler.struct.attr.*;
-import org.jetbrains.java.decompiler.struct.consts.LinkConstant;
 import org.jetbrains.java.decompiler.struct.consts.PrimitiveConstant;
 import org.jetbrains.java.decompiler.struct.gen.FieldDescriptor;
 import org.jetbrains.java.decompiler.struct.gen.MethodDescriptor;
@@ -335,78 +333,6 @@ public class ClassWriter {
     DecompilerContext.getLogger().endWriteClass();
   }
 
-  private static boolean hasAnnotations(StructMethod mt) {
-    return mt.getAttribute(StructGeneralAttribute.ATTRIBUTE_RUNTIME_INVISIBLE_ANNOTATIONS) != null ||
-      mt.getAttribute(StructGeneralAttribute.ATTRIBUTE_RUNTIME_VISIBLE_ANNOTATIONS) != null;
-  }
-
-  private static boolean isHiddenRecordMethod(StructClass cl, StructMethod mt, RootStatement root) {
-    if (cl.getRecordComponents() == null) return false;
-    return isSyntheticRecordMethod(mt, root) || isDefaultRecordMethod(mt, root) ||
-      (mt.getName().equals("<init>") && !hasAnnotations(mt) && isDefaultRecordConstructor(cl, root));
-  }
-
-  private static Exprent getSimpleReturnValue(RootStatement root) {
-    Statement block = root.getFirst();
-    if (!(block instanceof BasicBlockStatement)) return null;
-    List<Exprent> exprents = block.getExprents();
-    if (exprents.isEmpty()) return null;
-    Exprent exit = exprents.get(0);
-    if (!(exit instanceof ExitExprent)) return null;
-    return ((ExitExprent) exit).getValue();
-  }
-
-  @SuppressWarnings("SpellCheckingInspection")
-  private static boolean isSyntheticRecordMethod(StructMethod mt, RootStatement root) {
-    String name = mt.getName(), descriptor = mt.getDescriptor();
-    if ((!name.equals("equals") || !descriptor.equals("(Ljava/lang/Object;)Z")) &&
-      (!name.equals("hashCode") || !descriptor.equals("()I")) &&
-      (!name.equals("toString") || !descriptor.equals("()Ljava/lang/String;"))) {
-      return false;
-    }
-    Exprent value = getSimpleReturnValue(root);
-    if (!(value instanceof InvocationExprent)) return false;
-    LinkConstant bootstrapMethod = ((InvocationExprent) value).getBootstrapMethod();
-    if (bootstrapMethod == null) return false;
-    return "java/lang/runtime/ObjectMethods".equals(bootstrapMethod.classname) && "bootstrap".equals(bootstrapMethod.elementname);
-  }
-
-  // Simple heuristic to check if a method is a default getter generated by a record.
-  private static boolean isDefaultRecordMethod(StructMethod mt, RootStatement root) {
-    Exprent value = getSimpleReturnValue(root);
-    if (!(value instanceof FieldExprent)) return false;
-    FieldExprent fieldExprent = ((FieldExprent) value);
-    Exprent instance = fieldExprent.getInstance();
-    if (!(instance instanceof VarExprent)) return false;
-    return ((VarExprent) instance).getIndex() == 0 && fieldExprent.getName().equals(mt.getName());
-  }
-
-  private static boolean isDefaultRecordConstructor(StructClass cl, RootStatement root) {
-    List<StructRecordComponent> components = cl.getRecordComponents();
-    if (components == null) return false;
-    Statement block = root.getFirst();
-    if (!(block instanceof BasicBlockStatement)) return false;
-    List<Exprent> exprents = block.getExprents();
-    if (exprents.size() != components.size()) return false;
-    int lastIndex = 0;
-    for (int i = 0; i < components.size(); i++) {
-      StructRecordComponent component = components.get(i);
-      Exprent assignment = exprents.get(i);
-      if (!(assignment instanceof AssignmentExprent)) return false;
-      Exprent left = ((AssignmentExprent) assignment).getLeft();
-      if (!(left instanceof FieldExprent)) return false;
-      if (!component.getName().equals(((FieldExprent) left).getName())) return false;
-      Exprent fieldInstance = ((FieldExprent) left).getInstance();
-      if (!(fieldInstance instanceof VarExprent) || ((VarExprent) fieldInstance).getIndex() != 0) return false;
-      Exprent right = ((AssignmentExprent) assignment).getRight();
-      if (!(right instanceof VarExprent)) return false;
-      int index = ((VarExprent) right).getIndex();
-      if (index <= lastIndex) return false;
-      lastIndex = index;
-    }
-    return true;
-  }
-
   public static void packageInfoToJava(StructClass cl, TextBuffer buffer) {
     appendAnnotations(buffer, 0, cl, -1);
 
@@ -610,14 +536,7 @@ public class ClassWriter {
 
     if (components != null) {
       buffer.append('(');
-      for (int i = 0; i < components.size(); i++) {
-        StructRecordComponent cd = components.get(i);
-        if (i > 0) {
-          buffer.append(", ");
-        }
-        boolean varArgComponent = i == components.size() - 1 && isVarArgRecord(cl);
-        recordComponentToJava(cd, buffer, varArgComponent);
-      }
+      RecordHelper.appendRecordComponents(buffer, cl, components);
       buffer.append(')');
     }
 
@@ -674,13 +593,6 @@ public class ClassWriter {
       }
     }
     return false;
-  }
-
-  private static boolean isVarArgRecord(StructClass cl) {
-    String canonicalConstructorDescriptor =
-      cl.getRecordComponents().stream().map(c -> c.getDescriptor()).collect(Collectors.joining("", "(", ")V"));
-    StructMethod init = cl.getMethod(CodeConstants.INIT_NAME, canonicalConstructorDescriptor);
-    return init != null && init.hasModifier(CodeConstants.ACC_VARARGS);
   }
 
   private void fieldToJava(ClassWrapper wrapper, StructClass cl, StructField fd, TextBuffer buffer, int indent, BytecodeMappingTracer tracer) {
@@ -763,27 +675,6 @@ public class ClassWriter {
       buffer.append(";").appendLineSeparator();
       tracer.incrementCurrentSourceLine();
     }
-  }
-
-  private static void recordComponentToJava(StructRecordComponent cd, TextBuffer buffer, boolean varArgComponent) {
-    appendAnnotations(buffer, -1, cd, TypeAnnotation.FIELD);
-
-    Map.Entry<VarType, GenericFieldDescriptor> fieldTypeData = getFieldTypeData(cd);
-    VarType fieldType = fieldTypeData.getKey();
-    GenericFieldDescriptor descriptor = fieldTypeData.getValue();
-
-    if (descriptor != null) {
-      buffer.append(ExprProcessor.getCastTypeName(varArgComponent ? descriptor.type.decreaseArrayDim() : descriptor.type));
-    }
-    else {
-      buffer.append(ExprProcessor.getCastTypeName(varArgComponent ? fieldType.decreaseArrayDim() : fieldType));
-    }
-    if (varArgComponent) {
-      buffer.append("...");
-    }
-    buffer.append(' ');
-
-    buffer.append(cd.getName());
   }
 
   private static void methodLambdaToJava(ClassNode lambdaNode,
@@ -1137,7 +1028,7 @@ public class ClassWriter {
             BytecodeMappingTracer codeTracer = new BytecodeMappingTracer(tracer.getCurrentSourceLine());
 
             // Avoid generating imports for ObjectMethods during root.toJava(...)
-            if (isHiddenRecordMethod(cl, mt, root)) {
+            if (RecordHelper.isHiddenRecordMethod(cl, mt, root)) {
               hideMethod = true;
             } else {
               TextBuffer code = root.toJava(indent + 1, codeTracer);
@@ -1284,14 +1175,14 @@ public class ClassWriter {
     buffer.appendIndent(indent).append(" */").appendLineSeparator();
   }
 
-  private static final StructGeneralAttribute.Key<?>[] ANNOTATION_ATTRIBUTES = {
+  static final StructGeneralAttribute.Key<?>[] ANNOTATION_ATTRIBUTES = {
     StructGeneralAttribute.ATTRIBUTE_RUNTIME_VISIBLE_ANNOTATIONS, StructGeneralAttribute.ATTRIBUTE_RUNTIME_INVISIBLE_ANNOTATIONS};
-  private static final StructGeneralAttribute.Key<?>[] PARAMETER_ANNOTATION_ATTRIBUTES = {
+  static final StructGeneralAttribute.Key<?>[] PARAMETER_ANNOTATION_ATTRIBUTES = {
     StructGeneralAttribute.ATTRIBUTE_RUNTIME_VISIBLE_PARAMETER_ANNOTATIONS, StructGeneralAttribute.ATTRIBUTE_RUNTIME_INVISIBLE_PARAMETER_ANNOTATIONS};
-  private static final StructGeneralAttribute.Key<?>[] TYPE_ANNOTATION_ATTRIBUTES = {
+  static final StructGeneralAttribute.Key<?>[] TYPE_ANNOTATION_ATTRIBUTES = {
     StructGeneralAttribute.ATTRIBUTE_RUNTIME_VISIBLE_TYPE_ANNOTATIONS, StructGeneralAttribute.ATTRIBUTE_RUNTIME_INVISIBLE_TYPE_ANNOTATIONS};
 
-  private static void appendAnnotations(TextBuffer buffer, int indent, StructMember mb, int targetType) {
+  static void appendAnnotations(TextBuffer buffer, int indent, StructMember mb, int targetType) {
     Set<String> filter = new HashSet<>();
 
     for (StructGeneralAttribute.Key<?> key : ANNOTATION_ATTRIBUTES) {
